shmr_fisher/pyproject.toml
[project]
name = "shmr-fisher"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "astropy>=7.2.0",
    "colossus>=1.3.14",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pyyaml>=6.0",
    "scipy>=1.17.1",
]
